pocketflow-tool-pdf-vision/requirements.txt
pocketflow>=0.1.0
openai>=1.0.0
PyMuPDF>=1.22.0  # for PDF processing
Pillow>=10.0.0   # for image processing
python-dotenv>=0.19.0
